+++ tests/chemistry/test_morgan_fingerprint.py
@@ -3,12 +3,15 @@ import importlib
 import pytest
 
 import janitor.chemistry  # noqa: disable=unused-import
+from helpers import running_on_ci
 
-
-@pytest.mark.skipif(
-    importlib.util.find_spec("rdkit") is None,
+# Skip all tests if rdkit not installed
+pytestmark = pytest.mark.skipif(
+    (importlib.util.find_spec("rdkit") is None) & ~running_on_ci(),
     reason="rdkit tests only required for CI",
 )
+
+
 @pytest.mark.chemistry
 def test_morgan_fingerprint_counts(chemdf):
     """Test counts of Morgan Fingerprints converted from Mol objects."""
@@ -19,10 +22,6 @@ def test_morgan_fingerprint_counts(chemdf):
     assert (morgans.to_numpy() >= 0).all()
 
 
-@pytest.mark.skipif(
-    importlib.util.find_spec("rdkit") is None,
-    reason="rdkit tests only required for CI",
-)
 @pytest.mark.chemistry
 def test_morgan_fingerprint_bits(chemdf):
     """Test bits of Morgan Fingerprints converted from Mol objects."""
@@ -33,10 +32,6 @@ def test_morgan_fingerprint_bits(chemdf):
     assert set(morgans.to_numpy().flatten().tolist()) == set([0, 1])
 
 
-@pytest.mark.skipif(
-    importlib.util.find_spec("rdkit") is None,
-    reason="rdkit tests only required for CI",
-)
 @pytest.mark.chemistry
 def test_morgan_fingerprint_kind_error(chemdf):
     """Test ``morgan_fingerprint`` raises exception for invalid ``kind``."""